refactor(row-bridge): promote build_decorated_key/build_row/encode_clustering to shared crate
The write-direction row encoder (DecoratedKey + Row assembly) lived in
ferrosa-cql/bridge.rs; the Postgres front-end needs the SAME encoding to write
rows the engine + CQL reads decode correctly. Per D10 (ferrosa-postgres must not
depend on the 54k-LOC ferrosa-cql), move the encoders into the shared
ferrosa-row-bridge crate — ONE encoder, no duplication (a divergent row encoder
is the top SQL-front-end FMEA risk).

- ferrosa-row-bridge: add build_decorated_key/build_row/encode_clustering (use
  its own codec::encode_value -> byte-identical; CqlError -> RowBridgeError).
- ferrosa-cql/bridge.rs: delete the local defs, re-export from ferrosa-row-bridge
  (the existing From<RowBridgeError> for CqlError makes callers' ? transparent);
  build_delete_row/build_clustering_key stay and use the re-exported encode_clustering.
Behaviour-identical: ferrosa-cql 943 lib tests pass (incl. the moved bridge
encoder tests), ferrosa-row-bridge builds, clippy/fmt clean.
